add table created by lua_bundle to namespace
The name is @@@. Therefore, the table doesn't pollute the namespace
unless the programmer deliberately uses _ENV["@@@"] for his own
values.

This facility is helpful for interoperability with luaproc: Every thread
could receive some or all values values from this table to access
functionality otherwise only available to the main thread.